bot/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anthropic==0.42.0
google-api-python-client==2.155.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httpx==0.28.1
